[analyzer] Tests for Cppcheck standard (--std) handling
Analyzer tests were missing for the cppcheck standard parameter (--std)
forwarding. The standard translation is also tested in case of a gnu**
standard is given.
Also added a small change in the business logic for handling the
standard parameter. Changing the if to an elif should make it more
robust against future refactors.