report-converter: Parse all leaks reported by LeakSanitizer
Previously, report-converter was able to parse output by LeakSanitizer
only, if exactly one leak was found. This issue was reported as #3617.

By changing the regex to find the beginning of a single leak - instead
of finding the beginning of the entire LeakSanitizer report -
report-converter is now able to parse and convert LeakSanitizer output
with multiple leaks

The new line regex necessitates not skipping an empty line, which is
still required for 'old style' parsers, which can only parse one item
 - like ASan. For this reason the 'skip_line_after_match' attribute
has been introduced. When all sanitizer convers are able to parse multiple
defects, this attribute can be removed again.