[fix] Apply all fixits
Clang tidy may provide automatically applicable fixits. These can be
applied by clang-apply-replacements tool. This tool gets a directory as
input which contains .yaml files describing the replacements. If any of
these .yaml files can't apply for some reason, then the rest is skipped.

This clang tool is added a flag named --ignore-insert-conflict which
ignores the error event and continues the application of further .yaml
files. If CodeChecker recognizes the existence of this flag then it will
be used, otherwise CodeChecker creates a temporary directory with a
single .yaml file and applies them one by one.