[fix] Speed up resolved diffing
This commit tries to address the everlasting diff runtime, when the
report count is large (~60000) and the ReviewStatusRule count is also
substantial.
Two main edits were made, one is in getLinesInSourceFileContents
the other is in getReviewStatusRules api endpoints.
In getLinesInSourceFileContents my goal was to combine the requests into
chunks.
In getReviewStatusRules the goal was also the same, but here a
monolithic request was chopped into chunks.