[server] Update review status of reports in batch
Previously for every new inserted report, a review status query was done
to determine its target value to be set. This was factored out, and
instead is done in a single batch, at the end of inserting reports.

[server] Cache severity levels

Remove profiling code, and fix linter errors and tests

Refactor report storing, runtime performace gains

To be able to move the session.flush() from after each report to after
the completion of the processing of all report files, a new helper
method was needed to be introduced: __add_report_context().

This will handle the addition of notes, macro expansions in a separate
step, after the reports have been flushed to the db, and already gotten
their primary keys assigned.

Also added a helper class to handle review statuses stored in source
code.