[refactor] Analyzer context as singleton
`analyzer_context.Context` is a singleton class. Anywhere it is
constructed, the same object is acquired. So there is no need to pass
it as parameter through long function call chains, because this object
can be constructed where it's needed. It has no cost, because the same
instance is used once it was created.