[cmd] Fix a crash with CodeChecker cmd diff --unique on
Suppose we have a server with two non-identical runs, Example1 and
Example2. The following invocation:

CodeChecker cmd diff -o rows --url localhost/Default -n "Example2" -b "Example1" --new --uniqueing on

would have crashed at a thrift call level. If uniqueing wasn't on, the
crash was abscent as well.

The root of the problem wasn't on the thrift level, however, but rather
that a thrift function was invoked with None values, when said
parameters mustn't have been None.

Specifically, these values were the file ID's of some reports. The file
ID, column and row numbers of a report are None in the ReportData with
uniqueing, but are proper values otherwise.

This makes some sense -- if 10 reports have the same report hash, by
uniquing we obtain a report that describes 10 reports, so "some sense of
identity being lost" would be an understandable approach to uniquing,
but we can do better.

We can just take the arguably easier solution of picking a random report
out of the 10 equivalent one, and say that this report stands in the
place of the other 9.

On the implementation level, I simply querried the rest of the fields of
the table, which in effect does just this.