[bugfix] Checker config option format
"CodeChecker analyze" command has a --checker-config flag. The parameter
of this flag should be in the following format:
<analyzer>:<checker>:<option>=<value>

clang-tidy tool requires a "." character between the checker and the
option, so CodeChecker should convert it from : to .

The format of --checker-config (and --analyzer-config) is now checked
at the command line parsing phase, so it is done with a common algorithm
for all analyzer tools.

Earlier the default value of --analyzer-config was
"clang-tidy:HeaderFilterRegex='.*'", so clang-tidy analysis of header
files is also done. This is not a default value for the flag anymore,
but it is handled under clang-tidy module.