Rework store time logging
Now there is a separate logger. The handler for this new logger is
dynamically added to the current log configuration during log setup.
The former STORE_TIME log level (introduced in the previous commit)
could be eliminated, as well as the filter function for the handler.

Different severity levels can be logged against this new store_time
logger, (will be implemented in next commit after discussion):

* INFO - when there is no problem
* WARNING - when the store time gets over a threshold
* ERROR - when the store is not finished.

The report count has also been reworked, now a the MassStoreRun class
accounts this info in a member variable.