CodeChecker server hotswapper
The newly added small script watches for filesystem events inside the
web folder in this repo. If a suitable event occurs, first it stops the
Previous server, the starts a new one.
This feature depends on the dev_package target, it will only work with
that kind of CodeChecker build.