[bugfix] Fix the zombie process issue
Fixes #3442, where if the SIGINT signal was sent to CodeChecker during a
mulithread analysis, zombie processes were left. The problem was that
the signal handler actually threw an exception:

free variable 'pool' referenced before assignment in enclosing scope

This patch fixes this by moving signal.signal below the creation of the
pool variable.

Also, there is this site I randomly opened that said

"The process pool may take a moment to forcefully kill all child
processes, so it is a good practice to call the join() function after
calling terminate()."
https://superfastpython.com/multiprocessing-pool-kill-all-tasks/

so I did that, though thats just blindly listening to the article.