[feat] Support report annotations and add dynamic analyzer related annotations
A plist file may contains a section for report annotations:

```.xml
<key>diagnostics</key>
<array>
  <dict>
    <key>category</key>
    <string>Memory error</string>
    ...
    <dict>
      <key>timestamp</key>
      <string>2000-01-01 10:00</string>
      <key>testsuite</key>
      <string>TS-1</key>
      ...
    </dict>
  </dict>
</array>
```

At the time of creation of this commit there is no analyzer that would
emit a .plist file with such content. But the plan is that sanitizers and
some other dynamic analyzers will provide this information, at least the
report converter could add the timestamp to the .plist files.

The goal of this new feature is to enable storing arbitrary annotations to
the reports. We could have extended the report table with further columns
for each annotation, or we could have joined them in separate tables. Finally
we decided to create a generic "report_annotations" table where the
annotations are collected as (report_id; key; value) tuples. This adds some
complexity when selecting annotations for each report, but this tradeoff is
accepted.

Annotations are untyped labels, but "timestamp" and "testsuite" will be
handled in a specific way in the sense that users can order reports by their
timestamps and can filter reports based on which testsuite resulted them.
This can be useful because in case of dynamic analyzers it is possible that
a report is a consequence of another one. Ordering them by their timestamps
can help in detecting these relationships.

This commit targets the addition of "timestamp" and "testsuite" annotations,
but the implementation is more general than supporting only these two.
This is the reason why the database schema and Thrift API interfaces don't
have the concept of dynamic analysis.