[fix] Treat clang-diagnostic-* checkers as compiler flags
Previously the handling of clang-diagnostics was flawed,
because tidy cannot enable those as a checker, only suppression
of the resulting reports are available:
https://github.com/kimgr/clang-tools-extra/blob/master/docs/clang-tidy.rst

After this patch all of these checkers will be treated as compiler warning
flags, this way the compiler warnings can be enabled as
clang-diagnostics.
There is an added benefit of this method, as the checker prefix based
activation is also working now.
The old -W notation will be deprecated, but for the time being still usable.