[report-converter] Enable multiple input for report-converter
Sometimes dynamic analysis is executed for the code base several times
with different parameters. The output of these analyses may go to
different log files which are given to report-converter one-by-one.
If there are reports for the same source file in different log output
files then the resulting .plist files will be overwritten at a
subsequent execution of report-converter.

In this patch we make it possible to give multiple log files or
directories containing log files to the report-converter. The reports
in these files and directories will be collected all and will be
converted to .plist without overriding the ones coming from a different
log file.