Makefile: package_gerrit_skiplist should depend on package_dir_structure
In the Makefile, the `package_gerrit_skiplist` target copies a script to
a directory created in the `package_dir_structure` target. However,
`package_gerrit_skiplist` does not have a dependency on `package_dir_structure`,
and it is possible that `package_gerrit_skiplist` is run before the directory
created in `package_dir_structure` actually exists. This can be reproduced by
running `make` with multiple jobs.