[ldlogger] Recognize linux_spawn alongside exec* calls
The simple idea behind ldlogger is the following -- create a .so file
that we force to be loaded for each program that is executed by adding
it to the LD_PRELOAD environmental variable. In that .so, we shadow
exec* calls with our impostor exec* calls -- instead of the real thing,
our version is called, which logs the compile command, and then calls
the actual system function.

Up to, but not including GNU Make 4.3, only exec* system calls were
used. However, as of 4.3, posix_spawn is used by default, so this patch
adds shadow functions for posix_spawn and the very similar posix_spawnp.