[feat] Exclude dynamic parts of checker message in hash generation
Some dynamic analyzers, like sanitizers generate dynamic parts in the
checker message (e.g. memory addresses). These are changing in every
execution of the analyer. Since the checker message is part of the
report hash, the hash changes too. This is bad, because the report
itself is supposed to be the same.

The report converter can provide a static_message field for the report
which is a stable version of the checker message. If this field is
given then it will be used for hash generation.