Adapt to new clang-tidy checker options format.
Starting with LLVM15 / clang-tidy-15, the output of '--dump-config' has changed, and the old parser failed to parse the options in the new format. We need full YAML parsing now. This solution implements a simple trial-and-error approach: It first tries the old parsing version, and if that fails, it tries the new one. This seems a lot simpler than first trying to figure out the clang-tidy version by e.g. parsing the output of clang-tidy --version.