[gui] Update run filter when setting the filter options
Previously only the run history was affected, when the analyzed
before/after options were set. This resulted in a confusing view when
the displayed run entries last analyzed date would show outside of the
specified date.
This patch fixes this issue as it will hide those runs where the
before/after criteria is not met.