[fix] Refine when a report is regarded as outstanding for tags
Despite only two lines of code, this patch is many things all in once:
* We fix a bug where diffing tags returned unexpected (and percieved to
  be incorrect) results,
* We redefine what we expect from diffing tags,
* We redefine the expected heaviour around review status rules.

Previously, we regarded review status rules are a timeless property, but
what is even more true, we didn't really know what we expect from them
when it came to diffing tags. This lead to confusion on the developer
side and on the user side as well, and lead to whack-a-mole issues and
patches like #3675, that was more driven by what users expected from
this feature than a comprehensive plan.

This is okay -- the review status feature and the tag feature grew in
their own world, and nobody can be faulted for being on top of these
features having a very solid specifications right out of the gate. This
patch solved this issue.

From this point on, our stance is the following: when we diff runs, we
always check whether a report is outstanding _at the time of the query_,
and for diffing tags or timestamps, we check whether a report is
outstanding _at the time of the tag/timestamp_.

A user-facing documentation is written in #4006, and can be previewed
here:
https://github.com/Szelethus/codechecker/blob/diff_docs_rewrite/docs/web/diff.md