Change the selected bubble's highlight
To highlight selected bubble on the report page currently we use only a 2px dashed border. It is hard to detect the active chip section when there are many bubbles in one report file. The active bubble is more highlighted when its opacity is 100 and 70% for the others. The font-weight is bold for active and lighter for non actives. The current modifications emphasize the selected bubble more then the dashed border.