[diff] Fix a CLI remote-remote bug where we checked the detection status
Ideally, when we use CodeChecker cmd diff, we should return exactly what
the server API calls return. We obviously can't do that when either the
baseline or the newline result contains a local directory, since they
are not on the server. In that case, we need to parse the results, check
for review status rules on the server, etc.

However, when both operands _are_ on the server, there is no need to do
additional filtering, lets just trust that the server returns correct
results. If not, we should fix it the server.