[NFC] ReviewStatusHandler for source code comments
Currently there are 2 ways of handling review statuses for reports:
- In source code comments:
  // codechecker_suppress [checker_name] message...
- In the GUI:
  This sets a "review status rule" which means a relationship between
  bug hashes and review statuses. All reports with the same bug hash get
  the same review status.

We are planning a 3rd way of setting review statuses:
  A review_status.yaml config file which contains review status mappings
  to files or directories.

The long-term plan is to create a ReviewStatusHandler class that handles
the review status of a report, in all three scenarios above.

This commit is just a refactoring which moves the 1st scenario-related
code (review status from source code comment) to the ReviewStatusHandler
class.