feat(store): Explicitly time the client out if the connection hung
This is a partial solution that attempts to gracefully kill an ongoing
`CodeChecker store` client if the reply to the `massStoreRun()` API call
does not arrive in a timely manner. After reaching the timeout, the
client throws an exception onto itself and presents an understandable
error message about the situation. This implementation is temporary,
thus the timeout is a hard-coded limit.

The client still exits with a failure in these cases, because there are
no guarantees (and no deterministic ways to check) whether the timeout
happened because the TCP connection failed and stuck (which is observed
in a lot of cases, unfortunately) or whether the server indeed took more
time than allotted to process the stored reports. A successful exit
signal would result in CI scripts thinking the store was successfull and
then potentially breaking later, unreliably, on the fact that the stored
run does not in fact exist, or is in an outdated state, etc.