[gcc][fix] Find the correct sarif file location
The wording in the [GCC
docs](https://gcc.gnu.org/onlinedocs/gcc-13.1.0/gcc/Diagnostic-Message-Formatting-Options.html)
is a little confusing, it says

"The ‘sarif-stderr’ and ‘sarif-file’ formats both emit diagnostics in SARIF
Version 2.1.0 format, either to stderr, or to a file named source.sarif,
respectively."

I thought this will create a sarif file right next to the source file,
but in reality, the file is created in the `directory` field of the
compilation database file. This patch fixed this.