chore(store): Give a name to the timeout watchdog thread
In some cases, spurious exceptions may crop up in the outputs when
something goes wrong, and in those cases, threads are usually just given
an automatic name like `Thread-1` or so. This patch adds a unique name
to the store watchdog so it's clear if that causes an exception (it
shouldn't).