[fix] Bugfix at --enable-all and disable warning
There was a bug in CodeChecker:

CodeChecker analyze --enable-all --disable clang-diagnostic-unused-variable

This command didn't disable that warning. The problem is that we
represent --enable-all with -Weverything in the clang-tidy analyzer
command, but -Wno-unused-variable has to be appended too. It was a
false assumption that
"clang-tidy -checks='-clang-diagnostic-unused-variable' main.c -- -Weverything"
does the job.