Fix multiprocess store on OSX
Even though we swapped a higher level interface to a lower level
(ProcessPoolExecutor -> Multiprocess.Pool), it is worth it because on
all 3 systems the multiprocess handling is the same.
Multiprocessing is broken on osx, (probably because of the fork -> spawn
change.) This patch intends to correct it by using the newly added
multiprocess library.
On Windows only single threaded execution is supported. With the
introduction of the multiprocess lib multiprocessing is now available.