[fix] fix venv_dev and multiprocess lib usage
Many functions and members of `multiprocess` lib are exposed
dynamically. For this reason pylint is not able to tell what
members this lib has (e.g. `multiprocess.Value()`. So these
pylint reports have been suppressed.

Another purpose of this commit is to make "superset" relationship
for venv_dev: if a CodeChecker developer installs venv_dev, then
the content of venv should always be installed. venv_dev just
extends them. `multiprocess` was a counterexample for this relationship:
it was installed only by `venv` and not `venv_dev`. Such a
mistake in the organization of requirements.txt files results
missing libraries for developers.