[GUI] Update report sidebar structure and add review status reactivity
This PR gives a new report sidebar structure on the report-detail page. There are two main types in the report tree. One of them is the Outstanding category that contains outstanding reports classified into severity groups. The another category is the Closed that contains resolved, and not resolved but closed reports such as false positive, intentional.

Opening the a report-detail page, the report sidebar automatically opens a specific tree where the current report is located. If we changed the review status, the sidebar would reacts immediately.