[doc] Hide "ignore" feature
There is a use-case where the "ignore" feature doesn't work intuitively:
Review status config file is identified through a symlink in the report
directory. The reason is that changes in the review status config file
should apply even if it was modified after analysis. If
"CodeChecker analyze" would copy the review status config file into the
report directory then local changes in the user's config file wouldn't
apply.

However, in case an "ignore" action is added after analysis, the reports
are already dumped to the .plist files and further commands like
"CodeChecker parse" don't consider ignored reports as closed.

This commit should be reverted when CodeChecker commands consider
ignored reports as closed. The feature itself is in the code, we just
don't show it to the users ;)