[report-converter] Allow for empty strings in sanitizer error msgs
Sometimes, memory addresses will simply be omitted from the sanitizer
error messages. In this case, ` .+ ` will not match the error message,
since this matches strings with the minimum length of 3, but the
address-less message usually looks like this:
`error on adress  , because`, with only 2 characters there. The fix is
simple -- simply change `.+` zo `.*`.

I also added generic messages to asan, because we were missing that.